File convert (#28)
* baby steps

* Still need to fix JSON and multple creates!

* more on conversion..

* file convert

* only need to fix the Pandoc Media path part

* Rdy for testing

* upload comment working

* file convert further

* added custom Pandoc install path to settings

* added [filename] parsing for image uploads

* added [filename] parsing for image uploads2

* file upload with only 1 action value

* supporting multiple file upload fields, canvas and signature mix

* Canvas working again

* Canvas working again, but not mulitple files. js needs adjustments

* multiple canvas upload working

* almost done with signature.. rendering multiples

* render multiple signatures done

* still need to tweak signature upload!!

* multiple signatures upload done

* start made with normal multiple file upload input tags

* Multiple input type of file support

* Adding Live image capture

* start on video screenshot

* Fixed files actions check, created Helper for email

* Fixed files actions check, created Helper for email

* made some changes to JSON instances as per request by Dennis

* instance styling

* instance styling

* instance styling tweaks

* instance styling tweaks added fonts

* style change for instance-style.css

* JSON schemas start

* Start on schemas and JSON rendering

* Start on schemas and JSON rendering 2

* added required check

* reordering json schema

* JSON schemas start

* reordering json schema

* reordering json schema

* JSON schemas start

* reordering json schema

* reordering json schema

* JSON schemas start

* JSON schemas start

* file action="convertfromdocx" added, instead of convert_from

* file action="convertfrom:docx" added, instead of convert_from

* Added separator argument for select, option and token

* readme update

* instances update and change to separators

* temporary removed capitalizing page titles

* Updated Title creation to support captitalLinks

* instance with hidden fields

* handle spaces in file name uploads

* fixed boolean mistake

* added parsing arguments for labels

* added parsing arguments for labels 2

* More file convert debug information

* pandoc image conversion name change

* removed extra filetype check with file uploads

* Changed mwfollow code. Added PROTO_RELATIVE/

* pandoc image conversion name change

* Added page create Title check before any other Title manipulations

* multiple creates for 1 page fixed (issues/123)

* changed check for valid title

* empty label for not rendered anymore

* FlexForm/pull/27

* added no_disable_on_submit

* phpList extension for FlexForm

* phpList extension for FlexForm 2

* Added uploaded file information to use in create or edit

* Added uploaded file information to use in create or edit 2

* file upload names available fixing.. Instable atm

* Added some class names to debug. Worked on aliasfields and on file names available in edit and create

* extension phplist rdy 2 eval

* extension phplist rdy 2 eval 2

* extension phplist rdy 2 eval 3

* v2.1 Readme update and release

---------

+++ src/Processors/Files/Canvas.php
@@ -57,35 +57,39 @@ class Canvas {
 	 * @throws \MWContentSerializationException
 	 * @throws \MWException
 	 */
-	public static function upload( string $canvasFile ) {
+	public static function upload( string $canvasFile, array $fileDetails ) {
 		global $wgUser;
-		$fields = Definitions::fileUploadFields();
 		if ( Config::isDebug() ) {
 			Debug::addToDebug( 'File upload start',
 							   [
-								   'fields' => $fields,
-								   'post'   => $_POST
+								   'file details' => $fileDetails,
+								   'post'   => $_POST,
+								   'canvasFile' => $canvasFile
 							   ] );
 		}
 
-		if ( $fields['target'] === false || $fields['target'] === '' ) {
+		$target = General::getJsonValue( 'wsform_file_target', $fileDetails );
+
+		if ( $target === false || $target === '' ) {
 			throw new FlexFormException(
 				wfMessage( 'flexform-fileupload-no-target' )->text(),
 				0
 			);
 		}
 
-		if ( $fields['pagecontent'] === false ) {
-			$fields['pagecontent'] = '';
+		$pageContent = General::getJsonValue( 'wsform_page_content', $fileDetails );
+
+		if ( $pageContent === false ) {
+			$pageContent = '';
 		}
 
-		$fields['comment'] = self::getSummary();
+		$comment = self::getSummary();
 
 		$filesCore = new FilesCore();
 		$uploadPath = $filesCore->getUploadDir();
 		$canvasFile = $filesCore->getBase64Data( $canvasFile );
 		$pageName = $filesCore->remove_extension_from_image(
-				General::makeUnderscoreFromSpace( $fields['target'] )
+				General::makeUnderscoreFromSpace( $target )
 			) . ".jpg";
 
 		$tmpFileName = uniqid() . '.jpg';
@@ -102,36 +106,41 @@ class Canvas {
 			);
 		}
 
-		$details = trim( $fields['pagecontent'] );
+		$details = trim( $pageContent );
 
-		if ( $fields['pagetemplate'] && $fields['parsecontent'] !== false ) {
-			$filePageTemplate = trim( $fields['pagetemplate'] );
+		$pageTemplate = General::getJsonValue( 'wsform_file_template', $fileDetails );
+		$parseContent = General::getJsonValue( 'wsform_parse_content', $fileDetails );
+
+		if ( $pageTemplate && $parseContent !== false ) {
+			$filePageTemplate = trim( $pageTemplate );
 			$details = ContentCore::setFileTemplate( $filePageTemplate, $details );
 		}
 
-		if ( $fields['parsecontent'] !== false ) {
+		if ( $parseContent !== false ) {
 			$details = ContentCore::parseTitle( $details );
 		}
 
 		// find any other form fields and put them into the title
 		$pageName = ContentCore::parseTitle( $pageName );
-		$uploadFile = new Upload();
-		$resultFileUpload = $uploadFile->uploadFileToWiki(
-			$uploadPath . $tmpFileName,
-			$pageName,
-			$wgUser,
-			$details,
-			$fields['comment'],
-			wfTimestampNow()
-		);
-		if ( $resultFileUpload !== true ) {
-			throw new FlexFormException(
-				$resultFileUpload,
-				0
+		if ( !Config::isDebug() ) {
+			$uploadFile       = new Upload();
+			$resultFileUpload = $uploadFile->uploadFileToWiki(
+				$uploadPath . $tmpFileName,
+				$pageName,
+				$wgUser,
+				$details,
+				$comment,
+				wfTimestampNow()
 			);
-		}
+			if ( $resultFileUpload !== true ) {
+				throw new FlexFormException(
+					$resultFileUpload,
+					0
+				);
+			}
 
-		unlink( $uploadPath . $tmpFileName );
+			unlink( $uploadPath . $tmpFileName );
+		}
 
 		return true;
 	}
